Fix issue with all zeroes random matrix

+++ matmul/matmul/src/matmul.c
@@ -1,7 +1,7 @@
 #include <omp.h>
 #include <stdio.h>
 #include <stdlib.h>
-#include <sys/time.h>
+#include <time.h> 
 
 const size_t N = 1024;
 
@@ -16,9 +16,9 @@ void ZeroMatrix(double* A, size_t N) {
 void RandomMatrix(double* A, size_t N) {
     srand(time(NULL));
 
-    for (int i = 0; i < N; i++) {
-        for (int j = 0; j < N; j++) {
-            A[i * N + j] = rand() / RAND_MAX;
+    for (size_t i = 0; i < N; i++) {
+        for (size_t j = 0; j < N; j++) {
+            A[i * N + j] = (double)rand() / RAND_MAX;
         }
     }
 }
